Pagination

+++ IDemotivator/Controllers/DemotivatorsController.cs
@@ -16,6 +16,8 @@ using System.IO;
 using System.Text.RegularExpressions;
 using IDemotivator.Filters;
 using Nest;
+using PagedList.Mvc;
+using PagedList;
 
 namespace IDemotivator.Controllers
 {
@@ -43,6 +45,15 @@ namespace IDemotivator.Controllers
             var demotivators = await db.Demotivators.Where(d => d.AspNetUserId == CurId).ToListAsync();
             return View(demotivators);
         }
+        [AllowAnonymous]
+        public async Task<ActionResult> ShowAll(int? page)
+        {
+            int pageSize = 5;
+            int pageNumber = (page ?? 1);
+            var demotivators = await db.Demotivators.ToListAsync();
+            return View(demotivators.ToPagedList(pageNumber, pageSize));
+        }
+
 
         // GET: Demotivators/Details/5
         [AllowAnonymous]
@@ -264,11 +275,21 @@ namespace IDemotivator.Controllers
             }
         }
 
+        public void DeleteLikes(int id)
+        {
+            var likes = db.Likes.Where(t => t.CommentId == id).ToList();
+            foreach(var item in likes)
+            {
+                db.Likes.Remove(item);
+            }
+        }
+
         public void DeleteComments(int id)
         {
             var comments = db.Comments.Where(t => t.DemotivatorId == id).ToList();
             foreach (var item in comments)
             {
+                DeleteLikes(item.Id);
                 db.Comments.Remove(item);
             }
         }
@@ -307,7 +328,13 @@ namespace IDemotivator.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            
+
             Demotivator demotivator = await db.Demotivators.FindAsync(id);
+            if(User.Identity.GetUserId() != demotivator.AspNetUserId)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             DeleteAdds(id);
             
             db.Demotivators.Remove(demotivator);
